Added both functional and manual calibration and arming
Currently motors only work when you run only one of them. have not tested manual code yet but i expect it to run similarly. When a motor beeps repeatedly forever, idk what that means. However, when a motor beeps repeatedly for about a couple seconds and stops and does it again, it means that it has not met the required throttle range. if i run motors 1 and 2, and i have motor 1 calibrated first, and then 2, motor 2 will beep repeatedly and motor 1 will work. and vise versa.