IR Startup
This should work. I believe this is how we should go about polling the code from the IR remote. The only problem faced here is how to extract the integer out of the code. I assume it is quite easy but I am unable to troubleshoot it at the moment.